+++ CHANGELOG.md
@@ -9,9 +9,11 @@ All notable changes to this project will be documented in this file.
 - Bump to `2024` Rust edition ([#52])
 - Replace `snafu` with `eyre` and `thiserror` for error handling ([#53])
   - Use `color-eyre` for prettier error reports
+- Use `tracing` for logging ([#54])
 
 [#52]: https://github.com/sbernauer/breakwater/pull/52
 [#53]: https://github.com/sbernauer/breakwater/pull/53
+[#54]: https://github.com/sbernauer/breakwater/pull/54
 
 ## [0.17.0] - 2025-02-14